refactor: add service facades to reduce component coupling
Create 4 facade modules in src/utils/ that re-export grouped service
APIs, replacing direct 3+ service imports in 5 UI components:

- terminal-services.js: shell-api + fs-api + terminal-api
  (used by TerminalPanel and BoardView)
- file-tree-api.js: fs-api + shell-api + clipboard-api
  (used by FileTree)
- skills-api-facade.js: skills-api + shell-api + dialog-api
  (used by SkillsView)
- tab-services.js: git-api + fs-api + config-api
  (used by TabManager)

Each component now imports from a single facade instead of 3 separate
service modules, reducing coupling without changing runtime behavior.

Closes #416
