+++ src/components/board-view.js
@@ -13,9 +13,7 @@ import {
   resolveCardStatus, findTabForTerminal, getTabNameForTerminal, computeFocusIndex,
   formatCardLabel,
 } from '../utils/board-helpers.js';
-import ptyApi from '../services/terminal-api.js';
-import shellApi from '../services/shell-api.js';
-import fsApi from '../services/fs-api.js';
+import { ptyApi, shellApi, fsApi } from '../utils/terminal-services.js';
 
 export class BoardView extends ComponentBase {
   constructor(container, tabManager) {
