+++ src/components/terminal-panel.js
@@ -15,9 +15,7 @@ import {
   splitTerminal,
   focusDirection as focusDirectionHelper,
 } from '../utils/terminal-subsystem.js';
-import shellApi from '../services/shell-api.js';
-import fsApi from '../services/fs-api.js';
-import ptyApi from '../services/terminal-api.js';
+import { shellApi, fsApi, ptyApi } from '../utils/terminal-services.js';
 
 export class TerminalPanel {
   constructor(container, cwd) {
